Table creation from source file partially complete
Creation of table using a source text file for words is partially done.
What remains should be the following:
- Generation of word key for use with struct node_value, and thus be
  able to write a proper representation to file.
  The generation algorithm should create the same list given the same
  source file and configuration.
- A function to write the list representation to file.

As the representation is now a dw_hashmap that is created will use a
hash of the string as key in dw_node, whereas a dw_hashmap that is read
from a file should have the word key as key in dw_node, making them
incompatible and thus requiring creation of another representation in
order for list creation and use in a single command.